added timer in result table

+++ app/Http/Controllers/ModuleController.php
@@ -82,6 +82,8 @@ class ModuleController extends Controller
         $module = Module::find($request->id);
 
         if (!$module) return response(['error' => 'Illegal Access'], 500);
+        if ($module->active) return response(['error' => 'Cannot update when its active.'], 500);
+
 
         $module->update([
             'name' => $request->name,
@@ -104,6 +106,7 @@ class ModuleController extends Controller
         $module = Module::find($request->id);
 
         if (!$module) return response(['error' => 'Illegal Access'], 500);
+        if ($module->active) return response(['error' => 'Cannot delate when its active.'], 500);
 
         $module->delete();
         return response([
